Fix typos in class names and remove couplings between unit tests.
The KP instance and the invalid session keys were shared between all the unit tests, so the execution of the error tests
was influenced by a previous execution of the "satisfactory" tests and viceversa.

Now, the KP instance will change on every test execution, and the invalid session keys will contain an UUID to prevent
collisions.